backend/requirements.txt
djangorestframework-simplejwt
django-rest-knox
